requirements.txt
# cffi==1.16.0
# logging==0.4.9.6
pillow==10.4.0
pillow_heif==0.18.0
pyheif==0.7.1
# pycparser==2.22
